Fix four QA-reported bugs and harden listener + RPC layer
QA bugs (manual testing on btcpay.zano.me, ClickUp 86c9hkcm5,
86c9j1kjx, 86c9j1wjq, 86c9kd647):

- Invoice stuck in Processing despite on-chain confirmations.
  ConfirmationsRequired treated unlock_time as a confirmation count
  but Zano's unlock_time is an absolute block height (or Unix
  timestamp >= 500_000_000). A normal payment with unlock_time =
  block_height + 10 demanded ~13 980 extra confirmations and
  invoices eventually flipped to Invalid. Now translates a
  height-form unlock_time into "extra confs needed", combines with
  the speed-policy/store-threshold base via Math.Max, and gates
  settlement on a separate IsTimestampLocked() check for the
  timestamp branch so locked outputs don't settle before the
  wall-clock unlock passes.

- ZANO missing from currency autocomplete. Plugin now contributes a
  CurrencyDataProvider so ZANO + each registered CA flow into
  CurrencyNameTable for divisibility-aware number formatting.
  (BTCPay's CurrenciesSuggestionsTagHelper still excludes Crypto=true
  entries from the datalist itself; full UX fix requires a separate
  upstream patch.)

- Zano logo offset in checkout QR centerpiece. zano.svg cropped from
  the wide 105x36 wordmark to a square 36x36 icon-only viewBox so
  the QR center renders without distortion.

- "Network Cost" buffer paid the network fee twice. PaymentMethodFee
  was 0.01 ZANO for native, which inflated the customer's amount due;
  the wallet then added its own fee on top so the merchant received
  invoice + 0.01 instead of the exact invoice amount. Now 0 for both
  native and CA — the sender's wallet pays the network fee from its
  own balance, merchant receives exactly the invoice amount.

Listener/RPC hardening (security audit):

- Same-asset subtransfer collapse. AggregateCandidates now picks the
  highest-height tier per (tx, pid, asset) and SUMS amounts within
  the tier so multi-output payments to one integrated address are
  recorded at full amount instead of one leg only. Extracted as a
  testable static helper.

- Stale-wallet confirmation aging removed. The old
  UpdateExistingPaymentConfirmations bumped confs from stored
  BlockHeight + daemon currentHeight whenever the candidate scan
  found nothing, silently aging reorged-out or wallet-dropped
  payments to Settled. Deleted; confirmations now update only when
  the wallet still reports the transaction. Confirmation count
  clamped to Math.Max(0, ...) so reorgs that briefly drop the daemon
  height never produce negatives.

- Paging cursor on get_recent_txs_and_info2. Hardcoded
  Offset=0/Count=200 missed payments past the first page on busy
  wallets (one wallet is shared between native ZANO and every
  registered CA). Now pages until every monitored payment_id has
  been matched, history exhausts, or a 10-page safety cap is hit.

- Signed long overflow on 18-decimal CA amounts. ZanoSubtransfer.
  Amount switched from long to decimal; an 18-decimal CA payment of
  10 units = 10^19 atomic exceeds long.MaxValue and used to crash
  JSON deserialization for the entire poll cycle. Listener candidate
  tuple, AggregateCandidates, and HandlePaymentData thread decimal
  end-to-end. New ZanoMoney.FromAtomic(decimal, int) overload (C#
  can't overload by return type, so distinct name).

- Sync-summary readiness. AllAvailable() now requires a non-empty
  configured-network set AND Synced && WalletAvailable per network
  — matches ZanoRpcProvider.IsAvailable. Previously only checked
  DaemonAvailable and returned true on empty Summaries, which made
  /api/v1/server/info report fullySynched=true during plugin
  startup. GetStatuses() now emits PaymentTypes.CHAIN.GetPaymentMethodId
  (canonical "ZANO-CHAIN") instead of the bare-cryptoCode
  PaymentMethodId.Parse output.

- Asset-id hex validation. ParseExtraAssets now rejects asset_ids
  whose 64 chars include anything outside [0-9a-f]; length-only check
  let typos register a CA whose listener could never match real
  wallet transfers.

- Controller route hardening. UIZanoStoreController POST now uses the
  validated route cryptoCode for blob.SetExcluded instead of the
  client-supplied viewModel.CryptoCode, closing a tampered-hidden-field
  vector that could toggle a different Zano asset than the one whose
  config was being saved.

- JsonRpcClient logging. ILogger<JsonRpcClient> injected via DI;
  deserialization failures log a structured 256-char preview instead
  of dumping raw wallet RPC bodies (payment ids, tx hashes, amounts,
  asset ids) to stdout.

Tests:

- 87/87 passing (was 59 before this batch).
- New regression coverage for: unlock_time as block height vs
  timestamp, multi-leg same-asset subtransfer summing, mempool->
  confirmed transition, amount > long.MaxValue, FromAtomic at 18
  decimals, asset_id hex/non-hex, controller route mismatch (covered
  by manual code review).

Deferred to follow-up tickets:

- Per-crypto SemaphoreSlim to serialize listener scans
  (https://app.clickup.com/t/86c9p1btm).
- Consolidated wallet polling for native + CAs sharing one wallet
  (https://app.clickup.com/t/86c9p1btp).
- Bounded RPC timeout/retry on JsonRpcClient
  (https://app.clickup.com/t/86c9p1btw).
- Drop-detection: flip Settled->Processing when wallet drops a tx
  (https://app.clickup.com/t/86c9p1bu0).
- Currency autocomplete UX fix (requires upstream patch or client-side
  injection).

+++ Plugins/Zano/RPC/Models/GetRecentTxsAndInfo2Response.cs
@@ -29,7 +29,13 @@ namespace BTCPayServer.Plugins.Zano.RPC.Models
 
     public class ZanoSubtransfer
     {
-        [JsonProperty("amount")] public long Amount { get; set; }
+        // Atomic-unit amounts are represented as `decimal` rather than signed `long` because
+        // a 64-bit signed type can't hold valid Confidential-Asset amounts above
+        // ~9.22 atomic units when Decimals=18 (ZanoAssetDefinition allows decimals up to 18,
+        // making 10 units = 10^19 atomic, which exceeds long.MaxValue and would crash the
+        // entire poll cycle on JSON deserialization). decimal carries up to ~7.9e28, so any
+        // realistic CA amount fits with room to spare.
+        [JsonProperty("amount")] public decimal Amount { get; set; }
         [JsonProperty("asset_id")] public string AssetId { get; set; }
         [JsonProperty("is_income")] public bool IsIncome { get; set; }
     }
